+++ src/main/java/site/travellaboratory/be/controller/auth/dto/pw/PwInquiryEmailRequest.java
@@ -1,4 +1,4 @@
-package site.travellaboratory.be.controller.auth.dto;
+package site.travellaboratory.be.controller.auth.dto.pw;
 
 public record PwInquiryEmailRequest(
     String username
